requirements-vercel.txt
# Vercel Requirements - Lightweight (No PyTorch/YOLO)
# ML inference handled by HuggingFace API

flask==3.1.0
python-dotenv==1.0.1
requests==2.31.0
opencv-python-headless==4.10.0.84
numpy>=1.26.2
Pillow==10.4.0
pymongo==4.10.1
dnspython==2.7.0
cloudinary==1.41.0
shapely>=2.1.2
